backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
pyodbc==5.0.1
sqlalchemy==2.0.23
